backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
apscheduler>=3.10.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
cryptography>=41.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
